+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### 9.4.13 - SNAPSHOT
   * (PW-1697) Fixed validation/parse pattern in field 29O
   * (PW-1697) MT306 changes in field 30I
+  * Added DistinguishedName with Builder in order to encapsulate the BIC branch name logic 
 
 #### 9.4.12 - November 2023
   * (PW-1697) Fixed validation pattern in fields 14[H,K,L,M,N,O] and 29J